% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleClozeHelpers.R
\name{add_clozeQ}
\alias{add_clozeQ}
\title{add_clozeQ}
\usage{
add_clozeQ(cl, ...)
}
\arguments{
\item{cl}{clozeL container object}

\item{...}{additional arguments to be passed to \code{\link{clozeQ}} for creation of the question object}
}
\value{
returns a \code{clozeL} question container object with one question added.
}
\description{
Add a question to a \code{\link{clozeL}} container object of questions
}
\seealso{
\code{\link{clozeQ}}
}
