% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleClozeHelpers.R
\name{add_num}
\alias{add_num}
\title{add_num}
\usage{
add_num(cl, answer, points = 1, pointfrac = 100, tol = 0.001)
}
\arguments{
\item{cl}{clozeL container object}

\item{answer}{the correct answer always given as the first options, additional answer options
might be added, separated by a wiggle symbol "~".}

\item{points}{maximum points for this question.}

\item{pointfrac}{Numeric vector of points fractions given for answer options. The first answer
must be correct and have 100, following options fraction between 0 and 100.}

\item{tol}{tolerance for numerical questions.}
}
\value{
returns a \code{clozeL} question container object with the num question added.
}
\description{
Add a fill-in-a-number question (exams package: "num", Moodle: "NUMERICAL")
to a \code{\link{clozeL}} container object of questions
}
\examples{
data('mtcars')
cl <- new_clozeL('Car quiz')
# Question: How many cylinders does a Ferrari Dino have?
add_num(cl, answer=6, points=1, tol=1)
# Question: How many mpg does a Datsun 710 make?
add_num(cl, answer=c(22.8,22.8), points=2, tol=c(1,4), pointfrac=c(100,50))
}
\seealso{
\code{\link{clozeQ}}
}
