% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleClozeHelpers.R
\name{add_schoice}
\alias{add_schoice}
\title{add_schoice}
\usage{
add_schoice(cl, options, correctno, points = 1, shuffle = FALSE)
}
\arguments{
\item{cl}{clozeL container object.}

\item{options}{MCQ answer options provided as a character vector.}

\item{correctno}{A single numerical, indicates the correct MCQ answer option.}

\item{points}{Numerical, the maximum points for this question.}

\item{shuffle}{Shuffle answer options randomly.}
}
\value{
returns a \code{clozeL} question container object with the schoice question added.
}
\description{
Add a single choice question (exams package: "schoice", Moodle: "MULTICHOICE") to a \code{\link{clozeL}} container object of questions.
}
\examples{
data('mtcars')
cl <- new_clozeL('Car quiz')
# Question: How many cylinders does a Datsun 710 have?
add_schoice(cl, options=sort(unique(mtcars$cyl)), correctno=1, points=1, shuffle=FALSE)

}
\seealso{
\code{\link{clozeQ}}
}
