% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleClozeHelpers.R
\name{add_string}
\alias{add_string}
\title{add_string}
\usage{
add_string(cl, answer, points = 1, pointfrac = 100)
}
\arguments{
\item{cl}{clozeL container object}

\item{answer}{the correct answer always given as the first options, additional answer options might be added, separated by a wiggle symbol "~".}

\item{points}{maximum points for this question.}

\item{pointfrac}{Numeric vector of points fractions given for answer options. The first answer must be correct and have 100, following options fraction between 0 and 100.}
}
\value{
returns a \code{clozeL} question container object with the num question added.
}
\description{
Add a fill-in-a-string question (exams package: "string", Moodle: "SHORTANSWER") to a \code{\link{clozeL}} container object of questions
}
\examples{
data('mtcars')
cl <- new_clozeL('Car quiz')
# Question: Which of the cars has the most horsepower?
add_string(cl, answer=c('Maserati Bora','Camaro Z28','Duster 360'), points=2, pointfrac=c(100,50,50))
}
\seealso{
\code{\link{clozeQ}}
}
