% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleClozeHelpers.R
\name{clozeL}
\alias{clozeL}
\title{clozeL}
\usage{
clozeL(name)
}
\arguments{
\item{name}{name of the cloze question for meta-information}
}
\value{
returns a \code{clozeL} question container object.
}
\description{
Create a \code{clozeL} container object to store the partial question \code{clozeQ} object for a cloze question
}
