% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleClozeHelpers.R
\name{clozeQ}
\alias{clozeQ}
\title{clozeQ}
\usage{
clozeQ(answer, type = "num", points = 1, pointfrac = 100,
  tol = 0.001, mcq_options = "", mcq_correctno = NULL,
  mcq_shuffle = FALSE, mchoice_display = "")
}
\arguments{
\item{answer}{the correct answer always given as the first options, additional answer options might be added, separated by a wiggle symbol "~". For \code{schoice} questions, this is ignored and the information from \code{mcq_correctno} is used.}

\item{type}{one of num, string, schoice}

\item{points}{maximum points for this question.}

\item{pointfrac}{points fraction given for answer options. For first answer, it must be 100, for following options a percentage between 0 and 100 may be provided, separated by "~".}

\item{tol}{tolerance for numerical questions.}

\item{mcq_options}{MCQ answer options provided as a character vector.}

\item{mcq_correctno}{Number indicating the correct MCQ answer option.}
}
\value{
returns a \code{clozeQ} question object.
}
\description{
Create an object for a single question as part of a cloze question
}
