% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleClozeHelpers.R
\name{extract_element}
\alias{extract_element}
\title{extract_element}
\usage{
extract_element(cl, elname)
}
\arguments{
\item{cl}{clozeL question container object}

\item{elname}{character with the name of the element to extract from each clozeQ object (e.g. \code{points})}
}
\value{
returns a vector which collects element \code{elname} from all \code{clozeQ} contained in \code{clozeL}.
}
\description{
Extract one element from all questions contained in the clozeL container object.
}
