% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleClozeHelpers.R
\name{new_clozeL}
\alias{new_clozeL}
\title{new_clozeL}
\usage{
new_clozeL(name)
}
\arguments{
\item{name}{A string, name of the cloze question for meta-information.}
}
\value{
returns a \code{clozeL} question container object.
}
\description{
Create a new \code{clozeL} container object to store the partial question \code{clozeQ} object for a cloze question.
}
\examples{
new_clozeL('Quiz 1')
}
