% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleClozeHelpers.R
\name{print_answerlist}
\alias{print_answerlist}
\title{print_answerlist}
\usage{
print_answerlist(cl)
}
\arguments{
\item{cl}{clozeL container object.}
}
\description{
Print the MCQ answerlist from the clozeL object as rmarkdown into the exam. Empty lines for non-MCQ questions are included as required by the exams package.
}
\examples{
data('mtcars')
cl <- new_clozeL('Car quiz')
# Question: How many cylinders does a Datsun 710 have?
add_schoice(cl, options=sort(unique(mtcars$cyl)), correctno=1, points=1, shuffle=FALSE)
print_answerlist()
}
