% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleClozeHelpers.R
\name{stdround}
\alias{stdround}
\title{stdround}
\usage{
stdround(x, digits = 3)
}
\arguments{
\item{x}{number to round}

\item{digits}{number of decimal places}
}
\value{
returns rounded number (numeric).
}
\description{
Standard "school" rounding as opposed to the rounding of the \code{round} function.
}
\examples{
stdround(4.255,2)
}
