% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moodleClozeHelpers.R
\name{write_regression_equation}
\alias{write_regression_equation}
\title{write_regression_equation}
\usage{
write_regression_equation(reg, digits = 3)
}
\arguments{
\item{reg}{regression object.}

\item{digits}{number of digits to round to.}
}
\value{
Latex string.
}
\description{
Write an estimated regression equation from a \code{lm} regression object as LaTeX string.
}
\examples{
data('mtcars')
m <- lm(mpg ~ disp + wt, data=mtcars)
write_regression_equation(m)
}
